requirements.txt
# Core ML frameworks (CUDA 11.8 compatible)
torch==2.1.1+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118

# Memory-efficient attention
xformers==0.0.23+cu118

# 8-bit optimizers
bitsandbytes==0.41.3.post2

# Diffusion & Transformers
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0

# LoRA & PEFT
peft==0.7.1

# Dataset & Image processing
datasets==2.15.0
Pillow==10.1.0

# Hugging Face Hub (specific version for compatibility)
huggingface-hub==0.19.4

# Training utilities
tqdm==4.67.1
tensorboard==2.15.1
safetensors==0.4.1

# Core utilities
numpy==1.26.4
scipy==1.11.4
fsspec==2023.10.0
filelock==3.20.0